+++ src/gicv2.rs
@@ -10,6 +10,14 @@ use self::registers::{Gicc, Gicd, GicdCtlr};
 use crate::{IntId, Trigger};
 use core::ptr::NonNull;
 use safe_mmio::{UniqueMmioPointer, field, field_shared};
+use thiserror::Error;
+
+/// An error which may be returned from operations on a GIC Redistributor.
+#[derive(Error, Debug, Clone, Copy, Eq, PartialEq)]
+pub enum Error {
+    #[error("Cannot enable interrupt {0:?}")]
+    CannotEnable(IntId),
+}
 
 /// Driver for an Arm Generic Interrupt Controller version 2.
 #[derive(Debug)]
@@ -56,7 +64,7 @@ impl GicV2<'_> {
     }
 
     /// Enables or disables the interrupt with the given ID.
-    pub fn enable_interrupt(&mut self, intid: IntId, enable: bool) -> Result<(), ()> {
+    pub fn enable_interrupt(&mut self, intid: IntId, enable: bool) -> Result<(), Error> {
         let index = (intid.0 / 32) as usize;
         let bit = 1 << (intid.0 % 32);
 
@@ -69,7 +77,7 @@ impl GicV2<'_> {
                 & bit)
                 == 0
             {
-                return Err(());
+                return Err(Error::CannotEnable(intid));
             }
         } else {
             field!(self.gicd, icenabler).get(index).unwrap().write(bit);